HAWNG-1855: Refactors the hawtiotest tests inside an internal package
* Prevents the test API being public in the project as no need to expose it

* test_functions.go
 * Adds a FindProjectRoot function that walks up the directory structure
   to locate go.mod so avoiding the need to add copious .. paths when
   locating the CRD files in the tests.
