+++ pkg/controller/internal/hawtiotest/kubernetes/controller_integration_kubernetes_test.go
@@ -12,7 +12,7 @@ import (
 
 	"k8s.io/apimachinery/pkg/types"
 
-	"github.com/hawtio/hawtio-operator/pkg/controller/hawtiotest"
+	"github.com/hawtio/hawtio-operator/pkg/controller/internal/hawtiotest"
 )
 
 var _ = Describe("Testing the Hawtio Controller", Ordered, func() {
