+++ pkg/controller/internal/hawtiotest/kubernetes/testdata/configmap.yml
