+++ pkg/controller/internal/hawtiotest/kubernetes/testdata/deployment.yml
