+++ pkg/controller/internal/hawtiotest/openshift/testdata/configmap.yml
