+++ pkg/controller/internal/hawtiotest/openshift/testdata/crds/clusterversions.config.openshift.io.yaml
