+++ pkg/controller/internal/hawtiotest/openshift/testdata/crds/consolelinks.console.openshift.io.yaml
