+++ pkg/controller/internal/hawtiotest/openshift/testdata/crds/consoleyamlsamples.console.openshift.io.yaml
