+++ pkg/controller/internal/hawtiotest/openshift/testdata/crds/oauthclients.oauth.openshift.io.yaml
