+++ pkg/controller/internal/hawtiotest/openshift/testdata/crds/routes.openshift.io.yaml
