+++ pkg/controller/internal/hawtiotest/openshift/testdata/deployment.yml
